+++ package.json
@@ -25,30 +25,28 @@
     "qwik": "qwik"
   },
   "devDependencies": {
-    "@builder.io/qwik": "^1.2.17",
     "@builder.io/qwik-city": "^1.2.17",
+    "@builder.io/qwik": "^1.2.17",
     "@types/eslint": "^8.44.4",
+    "@types/estree-jsx": "^1.0.3",
     "@types/node": "^20.8.4",
     "@typescript-eslint/eslint-plugin": "^6.7.5",
     "@typescript-eslint/parser": "^6.7.5",
     "autoprefixer": "^10.4.14",
-    "eslint": "^8.51.0",
     "eslint-plugin-qwik": "1.2.13",
+    "eslint": "^8.51.0",
     "postcss": "^8.4.31",
-    "prettier": "^3.0.3",
     "prettier-plugin-tailwindcss": "^0.5.4",
+    "prettier": "^3.0.3",
     "rehype-pretty-code": "^0.10.2",
     "sass": "^1.69.5",
+    "shiki": "^0.14.5",
     "tailwindcss": "3.3.3",
     "typescript": "^5.2.2",
     "undici": "^5.26.0",
+    "unified": "^10.1.2",
     "unist-util-visit": "^5.0.0",
-    "vite": "^4.4.11",
-    "vite-tsconfig-paths": "^4.2.1"
-  },
-  "dependencies": {
-    "@types/estree-jsx": "^1.0.3",
-    "shiki": "^0.14.5",
-    "unified": "^10.1.2"
+    "vite-tsconfig-paths": "^4.2.1",
+    "vite": "^4.4.11"
   }
 }
