+++ src/routes/layout.tsx
@@ -10,8 +10,8 @@ import { type RequestHandler } from "@builder.io/qwik-city";
 import { Aside } from "~/components/Aside/Aside";
 import { Footer } from "~/components/Footer/Footer";
 import { Header } from "~/components/Header/Header";
+import { components } from "~/components/MdxComponents/MdxComponents";
 import { Toc } from "~/components/Toc/Toc";
-import { components } from "~/components/mdx-components";
 import { MDXProvider } from "~/state/MDXProvider";
 
 export const onGet: RequestHandler = async ({ cacheControl }) => {
